requirements.txt
matplotlib==3.6.1
numpy==1.23.4
pandas==1.5.0
python-binance==1.0.16
ta==0.10.2
